Use CMAKE_CURRENT_LIST_DIR
CMAKE_CURRENT_LIST_DIR is better because it is set to the directory of the file being processed (max-pretarget.cmake), regardless of the path of the file including this cmake file.

A relative path based on CMAKE_CURRENT_SOURCE_DIR makes too many assumptions about how the project is structured.